TASK-2: Move alert functionality to separate service Add success feedback for valid Todo decriptions
